requirements.txt
asgiref==3.8.1
cloudinary==1.39.0
Django==4.2.11
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
pillow==11.2.1
sqlparse==0.5.3
djangorestframework-simplejwt==5.0.0
djoser==2.2.2
django-allauth==0.57.0
dj-rest-auth==5.0.2
python-dotenv==1.0.1
dj-database-url==2.1.0
requests==2.31.0

# Celery and async task management
celery==5.3.4
flower==2.0.1
redis==5.0.1
psutil==5.9.8
supervisor==4.2.5
django-celery-results==2.5.1
django-celery-beat==2.5.0

# Production dependencies
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0